Conflicts between the fastverse and other packages

This function lists all the conflicts among fastverse packages and between fastverse packages and other attached packages. It can also be used to check conflicts for any other attached packages.

fastverse_conflicts(pkg = fastverse_packages())

Arguments

pkg
character. A vector of packages to check conflicts for. The default is all fastverse packages.

Details

There are 2 internal conflict in the core fastverse which are not displayed by fastverse_conflicts():

collapse::funique masks kit::funique. If both packages are detached, collapse is attached after kit. In general, the collapse version is often faster on data frames and supports unique rows on selected columns, but only preserves the first-match appearance order of character data (unique values for numeric columns are determined using a sort-based method). An option sort = TRUE lets collapse::funique return sorted unique values (for both character and numeric data). The kit version is generally faster for vectors and sometimes for data frames, and also supports matrices. It is a full hash-table based implementation that always preserves the first-match appearance order of data.

matrixStats::count masks kit::count. The matrixStats version is more flexible, supporting restricted search and missing value removal. The kit version is nearly twice as fast.

See also

Examples

# Check conflicts between fastverse packages and all attached packages
fastverse_conflicts()

# Check conflicts among all attached packages
fastverse_conflicts(rm_stub(search()[-1], "package:"))

#> -- Conflicts ------------------------------------------ fastverse_conflicts() --
#> x methods::body<-()    masks base::body<-()
#> x methods::kronecker() masks base::kronecker()